Add input checks to update the display correctly
When a button is pressed the input is checked to make sure the display
works as expected for a calculator. So no 2 operators in a row (except
negative numbers), decimal points always have a number before and after,
no operators in the middle of numbers, etc.

I used a lot of helper functions to help make the code more
self-explanatory and organized the functions into sections to make them
more easy to find.