Add chaining and special treatment of answer
For example "9 + 2" gives "= 18" when press =, and gives "= 18 +" if
press + instead. Can also press + after pressing =. This allows the user
to continue the calculation by using the previous answer.

Added "=" to the display output of the calculation to indicate to the
user that the first number is an answer, while also indicating to the
program that the answer should be deleted if an operator is not pressed
to continue the calculation.