Fix divide by 0 display message editing bug
Originally you could edit the divide by zero message output. Now when
this message is on the display the entire display is cleared upon a
button press and the program continues as normal from there.

I also had to fix chaining on the error message. Originally if you tried
n divided by 0 and pressed an operator the message would appear with the
operator added at the end! Now it only displays the error message.