Set limit to output given to display
Limited the output string so that it mostly always fits on the display.
This creates a better user experience.

Round down to 5 decimal places if the result is too long, and if it is
still too long then convert to exponential format with 3 decimal places.

The decimal places are decided based on making sure the output all fits
on the display. I tried to preserve the accuracy of the answer as much
as I could within this constraint which is why I try rounding to 5
decimal places first.

Note that it is still possible for the output to be too long, for
example if there are so many zeros that the exponent causes the output
to be too long. This could be fixed by continuing to reduce accuracy but
eventually the answer would be worthless. So lets stop at 3 decimal
places and let the output overflow the display width if it has to.
Anyway the display has horizontal scroll!