Add keyboard input
Pressing the key corresponding to the UI button will edit the display
accordingly. This is for a better user experience.

x and * both give multiply, and Enter and = both give equals. This is to
give the user extra options for what they may believe is the correct
button to press and therefore give a better user experience.

The key event listener simply fires the correct button click event, so
the rest of the code is the same as before. I did have to change the
button IDs due to the reason below, so the code to insert new symbols to
the nextDisplayText in processButtonPress() uses the text content rather
than the ID now. I also had to update the if statements in
specialButtonPressed() because of the ID name changes.

The overall process required changing the HTML button IDs to make the
document.querySelector() function work. A Switch statement is used for
keys that do not give a nice string to search buttons with, for example
the ID button "b-*" for the * key press gives an error.